requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
PySocks==1.7.1
sqlparse==0.5.3
